Reach point of no return
