+++ lib/pages/courses.dart
@@ -1,5 +1,7 @@
+import 'package:coursefidence/utils/course_model.dart';
 import 'package:flutter/material.dart';
-import 'package:coursefidence/utils/courseDS.dart';
+import 'package:coursefidence/widgets/add_new_course_form.dart';
+import 'package:provider/provider.dart';
 
 class Courses extends StatelessWidget {
   // _openCourseMenu(context) {
@@ -24,33 +26,56 @@ class Courses extends StatelessWidget {
   //   );
   // }
 
-  final List<Widget> courses = [
-    ListTile(title: Text("CSC 201"), trailing: Text("89%")),
-  ];
+  List<Widget> generatedCourses(modelList) {
+    // List<Widget> gen = [];
+    // final List<dynamic> courses = model.forEach(
+    //   (i) => ListTile(title: Text(i.name), trailing: Text("${i.confidence}%")),
+    // ).toList();
 
-  List<Widget> generatedCourses() {
-    List<Widget> gen = [];
-
-    for (int i = 0; i < courses.length; i++) {
-      gen.add(courses[i]);
-      gen.add(Divider());
-    }
-    return gen;
+    // for (int i = 0; i < courses.length; i++) {
+    //   gen.add(courses[i]);
+    //   gen.add(Divider());
+    // }
+    // List<dynamic> dude = modelList.map(
+    //   (i) => ListTile(title: Text(i.name), trailing: Text("${i.confidence}%")),
+    // ).toList();
+    // return dude;
   }
 
   @override
   Widget build(BuildContext context) {
+    final courseModel = Provider.of<CourseModel>(context);
+
     return Scaffold(
       appBar: AppBar(
         title: Text("Courses"),
       ),
       body: Container(
         alignment: Alignment.center,
-        child: ListView(
-          children: generatedCourses(),
+        child: Consumer<CourseModel>(
+          builder: (context, model, child) {
+            return ListView.builder(
+                itemCount: model.courseList.length,
+                itemBuilder: (context, index) {
+                  return ListTile(
+                    title: Text(model.courseList[index].name),
+                    trailing: Text("${model.courseList[index].confidence}"),
+                  );
+                }
+                );
+            // ListView.builder(
+            //   itemCount: courseModel.courseList.length,
+            //   itemBuilder: (context, index) {
+            //     final name = courseModel.courseList[index].name;
+            //     final confidence = courseModel.courseList[index].confidence;
+            //     return ListTile(
+            //         title: Text(name), trailing: Text("$confidence%"));
+            //   },
+            // );
+          },
         ),
       ),
-      floatingActionButton: Fab(courses: courses),
+      floatingActionButton: Fab(),
     );
   }
 }
@@ -66,154 +91,11 @@ class Fab extends StatelessWidget {
         showModalBottomSheet(
           context: context,
           builder: (BuildContext context) {
-            return Container(child: CourseForm(courses: courses));
+            return Container(child: CourseForm());
           },
         );
       },
       child: Icon(Icons.add),
-    );
-  }
-}
-
-class CourseForm extends StatefulWidget {
-  final List<Widget> courses;
-  CourseForm({this.courses});
-  @override
-  CourseFormState createState() {
-    return CourseFormState();
-  }
-}
-
-class CourseFormState extends State<CourseForm> {
-  final _formKey = GlobalKey<FormState>();
-  // final Map<String, dynamic> courseData = {
-  //   "courseName": null,
-  //   "courseUnits": null,
-  //   "confidenceLevel": null
-  // };
-
-  final courseData = new CourseDS();
-
-  get theCourseData {
-    return courseData;
-  }
-
-  final focusCourseUnits = FocusNode();
-
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-      key: _formKey,
-      child: Container(
-        height: 400,
-        // clipBehavior: Clip.round,
-        padding: EdgeInsets.symmetric(horizontal: 40, vertical: 20),
-        child: Column(
-          children: <Widget>[
-            Text(
-              "Add course",
-              style: TextStyle(fontSize: 25),
-            ),
-            TextFormField(
-              decoration: const InputDecoration(
-                icon: Icon(Icons.book),
-                hintText: 'Course Name',
-                labelText: 'Name *',
-              ),
-              validator: (value) {
-                if (value.isEmpty) {
-                  return 'Please enter the course name';
-                }
-                return null;
-              },
-              onSaved: (String value) {
-                courseData.courseName = value;
-              },
-              onFieldSubmitted: (v) {
-                FocusScope.of(context).requestFocus(focusCourseUnits);
-              },
-            ),
-            TextFormField(
-              decoration: const InputDecoration(
-                icon: Icon(Icons.confirmation_number),
-                hintText: 'Course Units',
-                labelText: 'Units *',
-              ),
-              validator: (value) {
-                if (value.isEmpty) {
-                  return 'Please enter the course unit';
-                }
-                return null;
-              },
-              onSaved: (String value) {
-                courseData.courseUnits = value;
-              },
-              focusNode: focusCourseUnits,
-            ),
-            ConfidenceSlide(course: courseData),
-            SizedBox(height: 30),
-            Builder(
-              builder: (BuildContext context) {
-                return RaisedButton(
-                  onPressed: () {
-                    if (_formKey.currentState.validate()) {
-                      _formKey.currentState.save();
-                      setState(() {
-                        widget.courses.add(
-                          ListTile(
-                            title: Text("${courseData.name}"),
-                            trailing: Text("${courseData.confidence}%"),
-                          ),
-                        );
-                      });
-                    }
-                  },
-                  child: Text('Submit'),
-                );
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class ConfidenceSlide extends StatefulWidget {
-  final CourseDS course;
-
-  ConfidenceSlide({this.course});
-  @override
-  SlideState createState() => SlideState();
-}
-
-class SlideState extends State<ConfidenceSlide> {
-  var _duelCommandment = 0;
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisSize: MainAxisSize.max,
-      children: <Widget>[
-        Text(
-          "Confidence level",
-          style: TextStyle(fontSize: 15, color: Colors.black54),
-        ),
-        Slider.adaptive(
-          value: _duelCommandment.toDouble(),
-          min: 0,
-          max: 100.0,
-          divisions: 99,
-          label: '$_duelCommandment%',
-          // onChanged: widget.onChangedSlide,
-          onChanged: (double newValue) {
-            // widget.onChangedSlide();
-            setState(() {
-              widget.course.confidence = newValue.round();
-              _duelCommandment = newValue.round();
-            });
-          },
-        ),
-      ],
     );
   }
 }
