+++ todo
@@ -1,2 +1,8 @@
-Implement the scroll selector
+Use a list view for the form
+Cause a UI rebuild when course is submitted
+Notify the user if he tries submitting the form without selecting a confidence level
+
+
 Implement the percent indicator
+
+
